send authtoken via query param
