+++ packages/@uppy/companion/src/server/controllers/index.js
@@ -1,6 +1,7 @@
 module.exports = {
   authorized: require('./authorized'),
   callback: require('./callback'),
+  sendToken: require('./send-token'),
   get: require('./get'),
   thumbnail: require('./thumbnail'),
   list: require('./list'),
